ajustes

+++ controllers/tareaController.js
@@ -15,7 +15,7 @@ exports.crearTarea = async (req, res) => {
     try {
 
         // Extraer el proyecto y comprobar si existe
-        const { proyecto } = req.query;
+        const { proyecto } = req.body;
 
         const existeProyecto = await Proyecto.findById(proyecto);
         if(!existeProyecto) {
@@ -40,82 +40,45 @@ exports.crearTarea = async (req, res) => {
 }
 
 // Obtiene las tareas por proyecto
-exports.obtenerTareas = async (req,res) => {
+exports.obtenerTareas = async (req, res) => {
 
-      try {
+        try {
+            // Extraer el proyecto y comprobar si existe
+            const { proyecto } = req.query;
 
-          // Extraer el proyecto y comprobar si existe
-          const { proyecto } = req.body;
 
-          const existeProyecto = await Proyecto.findById(proyecto);
-          if(!existeProyecto) {
-              return res.status(404).json({msg: 'Proyecto no encontrado'})
-          }
-  
-          // Revisar si el proyecto actual pertenece al usuario autenticado
-          if(existeProyecto.creador.toString() !== req.usuario.id ) {
-              return res.status(401).json({msg: 'No Autorizado'});
-          }
-          
-          // Obtener las tareas por proyecto
-          const tareas = await Tarea.find({proyecto}).sort({creado: -1});
-          res.json({tareas});
+            const existeProyecto = await Proyecto.findById(proyecto);
+            if(!existeProyecto) {
+                return res.status(404).json({msg: 'Proyecto no encontrado'})
+            }
 
-      } catch (error) {
-          console.log(error);
-          res.status(500).send('Hubo un error');
-      }
+            // Revisar si el proyecto actual pertenece al usuario autenticado
+            if(existeProyecto.creador.toString() !== req.usuario.id ) {
+                return res.status(401).json({msg: 'No Autorizado'});
+            }
+
+            // Obtener las tareas por proyecto
+            const tareas = await Tarea.find({ proyecto }).sort({ creado: -1 });
+            res.json({ tareas });
+
+        } catch (error) {
+            console.log(error);
+            res.status(500).send('Hubo un error');
+        }
 }
 
-// actualizar una tarea
-exports.actualizarTarea = async (req, res) => {
-    try {
-          // Extraer el proyecto y comprobar si existe
-          const { proyecto, nombre, estado   } = req.body;
-
-          // Si la tarea existe o no   
-          let tarea = await Tarea.findById(req.params.id);
-
-          if(!tarea){
-              return res.status(404).json({msg: 'Tarea no encontrada'});
-          }
-
-          // extraer proyecto
-          const existeProyecto = await Proyecto.findById(proyecto);
-  
-          // Revisar si el proyecto actual pertenece al usuario autenticado
-          if(existeProyecto.creador.toString() !== req.usuario.id ) {
-              return res.status(401).json({msg: 'No Autorizado'});
-          }        
-        
-          // Crear un nuevo objeto con la nueva información
-          const nuevaTarea = {};
-            nuevaTarea.nombre = nombre;
-            nuevaTarea.estado = estado;
-        
-
-        // Guardar la tarea
-        tarea = await Tarea.findOneAndUpdate({_id: req.params.id}, nuevaTarea, { new: true});
-
-        res.json({tarea});
-
-    } catch (error) {
-        console.log(error);
-        res.status(500).send('Hubo un error');
-    }
-}
-
-// elimina una tarea
-exports.eliminarTarea = async (req, res) => {
+// Actualizar una tarea
+exports.actualizarTarea = async (req, res ) => {
     try {
         // Extraer el proyecto y comprobar si existe
-        const { proyecto} = req.query;
+        const { proyecto, nombre, estado } = req.body;
 
-        // Si la tarea existe o no   
+
+        // Si la tarea existe o no
         let tarea = await Tarea.findById(req.params.id);
 
-        if(!tarea){
-            return res.status(404).json({msg: 'Tarea no encontrada'});
+        if(!tarea) {
+            return res.status(404).json({msg: 'No existe esa tarea'});
         }
 
         // extraer proyecto
@@ -124,14 +87,51 @@ exports.eliminarTarea = async (req, res) => {
         // Revisar si el proyecto actual pertenece al usuario autenticado
         if(existeProyecto.creador.toString() !== req.usuario.id ) {
             return res.status(401).json({msg: 'No Autorizado'});
-        }        
-      
-       // Eliminar 
-       await Tarea.findOneAndRemove({_id: req.params.id});
-       res.json({msg: 'Tarea Eliminada'})
+        }
+        // Crear un objeto con la nueva información
+        const nuevaTarea = {};
+        nuevaTarea.nombre = nombre;
+        nuevaTarea.estado = estado;
 
-  } catch (error) {
-      console.log(error);
-      res.status(500).send('Hubo un error');
-  }
+        // Guardar la tarea
+        tarea = await Tarea.findOneAndUpdate({_id : req.params.id }, nuevaTarea, { new: true } );
+
+        res.json({ tarea });
+
+    } catch (error) {
+        console.log(error);
+        res.status(500).send('Hubo un error')
+    }
+}
+
+
+// Elimina una tarea
+exports.eliminarTarea = async (req, res) => {
+    try {
+        // Extraer el proyecto y comprobar si existe
+        const { proyecto  } = req.query;
+
+        // Si la tarea existe o no
+        let tarea = await Tarea.findById(req.params.id);
+
+        if(!tarea) {
+            return res.status(404).json({msg: 'No existe esa tarea'});
+        }
+
+        // extraer proyecto
+        const existeProyecto = await Proyecto.findById(proyecto);
+
+        // Revisar si el proyecto actual pertenece al usuario autenticado
+        if(existeProyecto.creador.toString() !== req.usuario.id ) {
+            return res.status(401).json({msg: 'No Autorizado'});
+        }
+
+        // Eliminar
+        await Tarea.findOneAndRemove({_id: req.params.id});
+        res.json({msg: 'Tarea Eliminada'})
+
+    } catch (error) {
+        console.log(error);
+        res.status(500).send('Hubo un error')
+    }
 }
